Editing a Cabin

+++ src/services/apiCabins.js
@@ -11,21 +11,30 @@ export async function getCabins() {
   return data;
 }
 
-export async function createCabin(newCabin) {
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
+
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     '/',
     ''
   );
 
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
   // https://zkchcolbgtdeyjrrhetu.supabase.co/storage/v1/object/public/cabin-images/cabin-001.jpg
 
-  // 1. Create Cabin
-  const { data, error } = await supabase
-    .from('cabins')
-    .insert([{ ...newCabin, image: imagePath }])
-    .select();
+  // 1. Create/Edit Cabin
+  let query = supabase.from('cabins');
+
+  // A) Create
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+  // B) Edit
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq('id', id);
+
+  const { data, error } = await query.select().single();
 
   if (error) {
     console.error(error);
@@ -33,18 +42,19 @@ export async function createCabin(newCabin) {
   }
 
   // 2. Uplpad image
+  if (!hasImagePath) {
+    const { error: storageError } = await supabase.storage
+      .from('cabin-images')
+      .upload(imageName, newCabin.image);
 
-  const { error: storageError } = await supabase.storage
-    .from('cabin-images')
-    .upload(imageName, newCabin.image);
-
-  // 3. Delte the cabin If there was an error uploading image
-  if (storageError) {
-    await supabase.from('cabins').delete().eq('id', data.id);
-    console.error(storageError);
-    throw new Error(
-      'Cabin image cpulde not be uploaded and the cabin was not created'
-    );
+    // 3. Delte the cabin If there was an error uploading image
+    if (storageError) {
+      await supabase.from('cabins').delete().eq('id', data.id);
+      console.error(storageError);
+      throw new Error(
+        'Cabin image cpulde not be uploaded and the cabin was not created'
+      );
+    }
   }
 
   return data;
